Criado componentente separado para ser filho do NotificationMenu
